+++ src/app/ui/mod.rs
@@ -56,6 +56,7 @@ pub mod icon {
     pub const SUB: &str = "\u{2796}";
     pub const DISK: &str = "\u{1F5B4}";
     pub const ARROW_DOWN: &str = "\u{2B8B}";
+    pub const RESTART: &str = "\u{1F504}";
 }
 
 pub fn light_visuals() -> Visuals {